requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
asyncore-wsgi==0.0.10
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.7

click==8.1.3
colorama==0.4.4
colorlog==6.6.0
dateparser==1.1.0
epdb==0.15.1
filelock==3.7.1
Flask==2.2.2
frozenlist==1.2.0
idna==3.3
importlib-metadata==4.12.0
intervaltree==3.1.0
isal==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
lz4==3.1.3
MarkupSafe==2.1.1
multidict==5.2.0
mypy==0.910
mypy-extensions==0.4.3
networkx==2.8
numpy==1.23.3
pandas==1.4.2
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.2
requests==2.26.0
six==1.16.0
sortedcontainers==2.4.0
ta==0.10.1
TA-Lib==0.4.24
tag==0.5
tenacity==8.0.1
toml==0.10.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
udatetime==0.0.16
ujson==5.4.0
urllib3==1.26.7
websockets==9.1
Werkzeug==2.2.2
yarl==1.7.2
zipp==3.8.1
